Remove error page handling.
This software did provide an error page that was designed to help with
debugging the CGI script being executed. Although somewhat useful there
are a few reasons this needed to be removed:

* Security: It might expose sensitive information to the user. While
  this software is primarily for development purposes there might be valid
  reasons to use it in production. By providing that useful debugging
  information to users we are also providing it to potential hackers.
* Reduced Redundantancy: Rack provide Rack::ShowExceptions which provides very
  similar functionality. So we can just use that and provide almost the same
  info for debugging purposes.
* Flexibility: In addition to Rack::ShowExceptions there are other exception
  handling rack middleware that a project may want to use. By just throwning
  an exception any library can be used. Or the other middleware may capture
  the error and do something intelligent.
* Simplified Code: By removing our own error page we remove a lot of code. We
  also remove the temp file hack used to capture the standard out.
* Relevancy: I am looking to switch the PHP support off CGI and onto the
  built-in PHP server. The error page was too specific to CGI.

I debated if this change break the API. Obviously if the app assumes that
we handle errors and doesn't have some sort of exception capturing it could
be considered as an API breakage. But just about every Rack stack has some sort
of module of last resort to handle un-excepted exceptions. So overall I felt
it should not break any real world apps and therefore isn't an API breaking
release.