Dynamically determine port.
When switching to PHP's built-in webserver[1] I had to pick a port
for the server to run on. I picked 8080 as that is often a default
web development port. But since that could conflict with an existing
port in use I provided an option to specify the port.

I had a user report an issue[2] because they were running multiple instances
of Rack-Legacy causing a conflict on the port. Obviously they could just
use the option I provided but this prompted me to see if there was a way
to dynamically determine a free port.

I found reference[3] to a way using the TCPServer object. The user
reporting the issue did a quick test and found that it seemed to work
as advertised.

Since then it has been on my todo list to incorporate this auto-port
selection. I completely removed the port option as there is now no longer
any reason the user would need to select a port.

This commit is also taking this as an opportunity to change this project
to use symantic versioning. As this commit breaks the existing API (by
removing port option) I am assigning it the next major version number
(version 1). Also this software has been around long enough that it seems
incorrect to not even be at version one yet.

Most users are likely not specifying a port so in practice I don't think
this commit should break for many people. In fact for folks running multiple
instances of rack-legacy at once it might fix things.

This commit closes #20.

1. https://github.com/eric1234/rack-legacy/commit/14ae71bad2dfca52b365c90de042176a238d2340
2. https://github.com/eric1234/rack-legacy/issues/20
3. http://stackoverflow.com/a/201528/120067